+++ tools/validate/src/dependencies/validateListedDevDependencies.ts
@@ -3,8 +3,10 @@ import chalk from 'chalk';
 import pick from 'lodash/pick';
 import path from 'path';
 
-import { DepCheckFunctionProps } from '../config';
-import { ValidateCommandOptions } from '../validate.types';
+import {
+  DepCheckFunctionProps,
+  ValidateCommandOptions,
+} from '../validate.types';
 
 import { sortDependenciesByUsage } from './utils';
